requirements.txt
python-docx>=0.8.11
deep-translator>=1.11.4
openai>=1.44.0

